Repository maintenance: add GitHub Repo Gardening action
The Repo Gardening action is a tool that one can use to automate some of the tasks that you can perform to monitor activity in your repository.
https://github.com/marketplace/actions/repository-gardening

It includes different tasks, that can be enabled based on your needs.

This commit gets us started with the action by enabling a few tasks.

- This action relies on a few secrets that have already been added to this repo.
- Of note, `gatherSupportReferences` and `replyToCustomersReminder` aren't too useful right now, but will become useful once the plugin is released on WordPress.com and once we start receiving feedback from WordPress.com site owners.
- The `flagOss` task will point to #fediverse for now. This can be updated later on.
- I've added mapping for our 2 block directories, but that mapping can be extended for more automatic labeling later on.