Changelog: invite external contributors to add changelog file
Follow-up to #1479.

We cannot add a file to their branch automatically, unfortunately. We'll consequently take a different approach: if the PR is from a fork and the contributor provided changelog information in the PR description, we'll leave a comment on the PR inviting them to add the changelog file manually.