Remove "Followed" date
Now that we use the persistence for following and followers, we can not say for sure if that date is the date of following or followed or simply the date when they were imported through an @-reply.

This PR removes it as quick fix. If we want to have that information, we would need to have it in a separate PR, using a post-meta for example.