Add filter to disable local attachment storage
Introduces the `activitypub_store_attachments_locally` filter to allow
users and plugins to disable local storage of attachments from incoming
ActivityPub posts. This addresses feedback about providing an option for
users with limited webspace who may prefer not to store remote images
locally.

The filter is applied in both the add() and update() methods of the
Posts collection class, with full documentation on the first instance
and a reference comment on the second instance following WordPress
conventions.