Remove invalid activitypub from NodeInfo services arrays
ActivityPub should be listed in the protocols array, not in the
services.inbound or services.outbound arrays. The NodeInfo schema
only allows specific service names in these arrays, and 'activitypub'
is not one of them.

According to the NodeInfo 2.0/2.1 schemas, the services arrays should
only contain platform names like 'twitter', 'facebook', 'diaspora', etc.
ActivityPub is correctly listed in the protocols array.

Fixes #2600